#19553 Fix One2OneBidi not failing if inner flow completion wins the race with cancellation
